refactor(upstream): extract pool/ensure.rs lazy seeding + tool discovery (lab-kvji.12.6 step 12)
Move seed_lazy_upstreams, ensure_lazy_upstream_entry, ensure_tools_for_upstream(+
_with_connector), install_test_tools_for_upstream, lazy_connect_lock,
reprobe_tools_for_upstream, and replace_catalog_tools into ensure.rs + the 5
seed/ensure_tools tests. pub(super) FLIP (plan §2.1/§3.0): replace_catalog_tools
promoted to pub(super) — reprobe_upstream (probe.rs, step 13) calls it across the
boundary. 470 LOC (<500). Behavior unchanged. 1423 green.

Bead: lab-kvji.12.6
